Re-implemented previous functionality with the new database structure
