+++ data/accounts.json
@@ -1,0 +1,1 @@
+[{"username":"teacher","passwordHash":731,"userType":1,"id":293135},{"username":"student","passwordHash":3706,"userType":0,"id":344062}]
