+++ src/Util.js
@@ -9,7 +9,6 @@ function ReadFile(filepath)
 
 function WriteToFile(filepath, data)
 {
-    console.log("writing to: " + filepath);
     fs.writeFile(filepath, data, function(err)
     {
         if (err)
